+++ src/components/ES6/_global/js/fn/UixParallax.js
@@ -3,8 +3,9 @@
  *************************************
  * Parallax Effect
  *
- * @param  {Number} speed     - The speed of movement between elements.
- * @param  {JSON} bg          - Specify the background display. Default value: { enable: true, xPos: '50%' }
+ * @param  {Number} speed       - The speed of movement between elements.
+ * @param  {String} transition  - Transition time can simulate easing effect.
+ * @param  {JSON} bg            - Specify the background display. Default value: { enable: true, xPos: '50%' }
  * @return {Void}
  *
  *************************************
@@ -17,6 +18,7 @@
         // This is the easiest way to have default options.
         var settings = $.extend({
 			speed    : 0.25,
+            transition : 'all 0.4s cubic-bezier(0, 0, 0.34, 0.96) 0s',
 			bg       : { enable: true, xPos: '50%' }
         }, options );
  
@@ -63,16 +65,21 @@
 					st       = $this.offset().top - scrolled;
 				
 
-				
 				if ( bgEff ) {
 					//background parallax
 					TweenMax.set( $this, {
-						backgroundPosition: bgXpos + ' ' + ( 0 - ( st * speed ) ) + 'px'
+                        css:{ 
+                            'background-position': bgXpos + ' ' + ( 0 - ( st * speed ) ) + 'px',
+                            'transition': settings.transition
+                        }
 					});
 				} else {
 					//element parallax
 					TweenMax.set( $this, {
-						y: ( 0 - ( scrolled * speed ) )
+                        css:{ 
+                            'transform': 'matrix(1, 0, 0, 1, 0, '+( 0 - ( scrolled * speed ) )+')',
+                            'transition': settings.transition
+                        }
 					});